Start refactor but with some tests failing
