+++ cypress/support/config/services.js
@@ -77,7 +77,7 @@ export default {
           Cypress.env('APP_ENV') === 'live' || Cypress.env('APP_ENV') === 'test'
             ? undefined
             : '/amharic/articles/c123456abcdo',
-        smoke: true,
+        smoke: false,
       },
       frontPage: {
         path: Cypress.env('APP_ENV') === 'live' ? undefined : '/amharic',
@@ -88,7 +88,7 @@ export default {
           Cypress.env('APP_ENV') === 'live'
             ? undefined
             : '/amharic/bbc_amharic_radio/liveradio',
-        smoke: true,
+        smoke: false,
       },
     },
   },
@@ -101,14 +101,14 @@ export default {
           Cypress.env('APP_ENV') === 'live'
             ? '/arabic/articles/c8j91j2ljppo'
             : '/arabic/articles/c69dvq19k63o',
-        smoke: true,
+        smoke: false,
       },
       errorPage404: {
         path:
           Cypress.env('APP_ENV') === 'live' || Cypress.env('APP_ENV') === 'test'
             ? undefined
             : '/arabic/articles/c123456abcdo',
-        smoke: true,
+        smoke: false,
       },
       frontPage: {
         path: Cypress.env('APP_ENV') === 'live' ? undefined : '/arabic',
@@ -219,21 +219,21 @@ export default {
           Cypress.env('APP_ENV') === 'live'
             ? undefined
             : '/cymrufyw/articles/c06p32z9x2mo',
-        smoke: true,
+        smoke: false,
       },
       errorPage404: {
         path:
           Cypress.env('APP_ENV') === 'live' || Cypress.env('APP_ENV') === 'test'
             ? undefined
             : '/cymrufyw/articles/c123456abcdo',
-        smoke: true,
+        smoke: false,
       },
       frontPage: {
         path:
           Cypress.env('APP_ENV') === 'live' || Cypress.env('APP_ENV') === 'test'
             ? undefined
             : '/cymrufyw',
-        smoke: true,
+        smoke: false,
       },
       liveRadio: { path: undefined, smoke: false },
     },
@@ -372,7 +372,7 @@ export default {
           Cypress.env('APP_ENV') === 'live' || Cypress.env('APP_ENV') === 'test'
             ? undefined
             : '/igbo/articles/cxvxrj8tvppo',
-        smoke: true,
+        smoke: false,
       },
       frontPage: { path: '/igbo', smoke: true },
       liveRadio: { path: undefined, smoke: false },
@@ -405,7 +405,7 @@ export default {
           Cypress.env('APP_ENV') === 'live'
             ? undefined
             : '/indonesia/bbc_indonesian_radio/liveradio',
-        smoke: true,
+        smoke: false,
       },
     },
   },
@@ -418,18 +418,18 @@ export default {
           Cypress.env('APP_ENV') === 'live'
             ? undefined
             : '/japanese/articles/c693w95w0mko',
-        smoke: true,
+        smoke: false,
       },
       errorPage404: {
         path:
           Cypress.env('APP_ENV') === 'live' || Cypress.env('APP_ENV') === 'test'
             ? undefined
             : '/japanese/articles/c123456abcdo',
-        smoke: true,
+        smoke: false,
       },
       frontPage: {
         path: Cypress.env('APP_ENV') === 'live' ? undefined : '/japanese',
-        smoke: true,
+        smoke: false,
       },
       liveRadio: { path: undefined, smoke: false },
     },
@@ -461,7 +461,7 @@ export default {
           Cypress.env('APP_ENV') === 'live'
             ? undefined
             : '/korean/bbc_korean_radio/liveradio',
-        smoke: true,
+        smoke: false,
       },
     },
   },
@@ -555,21 +555,21 @@ export default {
           Cypress.env('APP_ENV') === 'live' || Cypress.env('APP_ENV') === 'test'
             ? undefined
             : '/naidheachdan/articles/c18q7nedn2ko',
-        smoke: true,
+        smoke: false,
       },
       errorPage404: {
         path:
           Cypress.env('APP_ENV') === 'live' || Cypress.env('APP_ENV') === 'test'
             ? undefined
             : '/naidheachdan/articles/c123456abcdo',
-        smoke: true,
+        smoke: false,
       },
       frontPage: {
         path:
           Cypress.env('APP_ENV') === 'live' || Cypress.env('APP_ENV') === 'test'
             ? undefined
             : '/naidheachdan',
-        smoke: true,
+        smoke: false,
       },
       liveRadio: { path: undefined, smoke: false },
     },
@@ -636,7 +636,7 @@ export default {
           Cypress.env('APP_ENV') === 'live'
             ? '/pashto/articles/c70970g2251o'
             : '/pashto/articles/cng0e8v85eko',
-        smoke: true,
+        smoke: false,
       },
       errorPage404: {
         path:
@@ -705,7 +705,7 @@ export default {
           Cypress.env('APP_ENV') === 'live' || Cypress.env('APP_ENV') === 'test'
             ? undefined
             : '/pidgin/articles/cxvxrj8tvppo',
-        smoke: true,
+        smoke: false,
       },
       frontPage: { path: '/pidgin', smoke: true },
       liveRadio: { path: undefined, smoke: false },
@@ -961,10 +961,10 @@ export default {
           Cypress.env('APP_ENV') === 'live' || Cypress.env('APP_ENV') === 'test'
             ? undefined
             : '/thai/articles/c123456abcdo',
-        smoke: true,
+        smoke: false,
       },
       frontPage: {
-        path: Cypress.env('APP_ENV') === 'live' ? undefined : '/thai',
+        path: '/thai',
         smoke: true,
       },
       liveRadio: { path: undefined, smoke: false },
@@ -997,7 +997,7 @@ export default {
           Cypress.env('APP_ENV') === 'live'
             ? undefined
             : '/tigrinya/bbc_tigrinya_radio/liveradio',
-        smoke: true,
+        smoke: false,
       },
     },
   },
@@ -1070,14 +1070,14 @@ export default {
           Cypress.env('APP_ENV') === 'live'
             ? '/urdu/articles/c4qg7qq63y6o'
             : '/urdu/articles/cx621klkm1ro',
-        smoke: true,
+        smoke: false,
       },
       errorPage404: {
         path:
           Cypress.env('APP_ENV') === 'live' || Cypress.env('APP_ENV') === 'test'
             ? undefined
             : '/urdu/articles/c123456abcdo',
-        smoke: true,
+        smoke: false,
       },
       frontPage: {
         path: Cypress.env('APP_ENV') === 'live' ? undefined : '/urdu',
@@ -1164,7 +1164,7 @@ export default {
           Cypress.env('APP_ENV') === 'live' || Cypress.env('APP_ENV') === 'test'
             ? undefined
             : '/yoruba/articles/cxvxrj8tvppo',
-        smoke: true,
+        smoke: false,
       },
       frontPage: { path: '/yoruba', smoke: true },
       liveRadio: { path: undefined, smoke: false },
