+++ cypress/support/helpers/runTestsForPage.js
@@ -1,9 +1,8 @@
 import config from '../config/services';
 import shouldSmokeTest from './shouldSmokeTest';
-import {
-  runCommonTests,
-  runCommonCanonicalTests,
-} from '../../integration/pages';
+import testsForAllPages from '../../integration/pages/testsForAllPages';
+import testsForAllAMPPages from '../../integration/pages/testsForAllAMPPages';
+import testsForAllCanonicalPages from '../../integration/pages/testsForAllCanonicalPages';
 
 const serviceHasPageType = (service, pageType) =>
   config[service].pageTypes[pageType].path !== undefined;
@@ -25,8 +24,8 @@ const runTestsForPage = (
           });
         });
 
-        runCommonTests({ service, pageType });
-        runCommonCanonicalTests({ service, pageType });
+        testsForAllPages({ service, pageType });
+        testsForAllCanonicalPages({ service, pageType });
         if (runTests) runTests({ service, pageType });
         if (runCanonicalTests) runCanonicalTests({ service, pageType });
       });
@@ -38,7 +37,8 @@ const runTestsForPage = (
           });
         });
 
-        runCommonTests({ service, pageType });
+        testsForAllPages({ service, pageType });
+        testsForAllAMPPages(pageType);
         if (runTests) runTests({ service, pageType });
         if (runAmpTests) runAmpTests({ service, pageType });
       });
